==> benchmark/notebooks/4_differential_abundance_testing.Rmd <==
---
title: 'Aggregate to protein-level abundances'
author:
  - name: "Tom Smith"
    affiliation: "Cambridge Centre for Proteomics"
date: "`r format(Sys.time(), '%d %B, %Y')`"
abstract: | 
  Here, we use limma to perform differential protein abundance testing and then
  compare the results with the different filtering thresholds.
output:
  pdf_document:
  html_notebook: default
geometry: margin=1in
fontfamily: mathpazo
fontsize: 11pt
---

```{r}
library(camprotR)
library(tidyverse)
library(limma)
library(MSnbase)

select <- dplyr::select
```


```{r}
prot_res <- readRDS('../results/prot_res.rds')
pep_res <- readRDS('../results/pep_res.rds')
expected <- readRDS('../results/expected.rds')
psm_res <- readRDS('../results/psm_res.rds')
```


```{r}
length(prot_res$`AGC: 2E5`)
length(pep_res$`AGC: 2E5`)
```


Define a function to run limma, with option to run 'mock' or real differential abundance testing.
```{r}
run_limma <- function(obj,
                      mock=FALSE,
                      merge_col='Master.Protein.Accessions'){
  if(mock){
    # Treat the first 4 tags as duplicates for two conditions
    extract_cols <- 1:4
    spike <- factor(rep(c('a','b'), c(2,2)))
    contrasts <- 'spikeb'
  } else {
    extract_cols <- 1:10
    spike <- factor(rep(c(1,2,6), c(4,3,3)))
    contrasts <- c('spike2', 'spike6')
  }
  
  contrasts2desc <- c('Mock', '2x vs 1x', '6x vs 1x')
  names(contrasts2desc) <- c('spikeb', 'spike2', 'spike6')
  
  results <- vector('list', length(contrasts))
  names(results) <- unname(contrasts2desc[contrasts])
  
  study.design <- model.matrix(~ spike)
  
  dat <- exprs(obj)[,extract_cols] %>% log(base=2)
  
  fit <- lmFit(dat, study.design)
  fit <- eBayes(fit, trend=TRUE)

  for(contrast in contrasts){

    limma.results <- topTable(fit, coef = contrast, n = Inf, confint=TRUE)
    limma.results$sigma <- fit$sigma[rownames(limma.results)]

    limma.results <- limma.results %>%
      merge(fData(obj)[,'species',drop=FALSE], by='row.names') %>%
      filter(species!='mixed') %>%
      mutate(binned_ave_expr=Hmisc::cut2(AveExpr, g=8),
             comparison=contrasts2desc[[contrast]])
    
    results[[contrasts2desc[[contrast]]]] <- limma.results
  }
  
  results <- results %>% do.call(what='rbind')

  results
}

```

Run limma 'mock' over all datasets

```{r}

datasets <- list('protein'=prot_res, 'peptide'=pep_res)

all_mock_limma_results <- datasets %>% names() %>% lapply(function(l){
  
  datasets[[l]] %>% names() %>% lapply(function(r){
    
     datasets[[l]][[r]] %>% names() %>% lapply(function(t){
      
        obj <- datasets[[l]][[r]][[t]]$agg
        
        results <- run_limma(obj, mock=TRUE) %>%
          mutate(level=l, run=r, threshold=t)
        
        results
      }) %>% do.call(what='rbind')
    }) %>% do.call(what='rbind')
  }) %>% do.call(what='rbind')
```



Run limma for all datasets
```{r}


all_limma_results <- datasets %>% names() %>% lapply(function(l){
  
  datasets[[l]] %>% names() %>% lapply(function(r){
    
     datasets[[l]][[r]] %>% names() %>% lapply(function(thresh){
      
        obj <- datasets[[l]][[r]][[thresh]]$agg
        results <- run_limma(obj) %>%
          mutate(level=l, run=r, threshold=thresh)
        
        results
      }) %>% do.call(what='rbind')
    }) %>% do.call(what='rbind')
  }) %>% do.call(what='rbind')
```



```{r}
runs <- unique(all_limma_results$run)
names(runs) <- runs

levels <- unique(all_limma_results$level)
names(levels) <- levels

```

```{r}
contains_notch <- psm_res %>% lapply(function(x){
  levels %>% lapply(function(level){
  
    if(level=='protein'){
      group_id <- 'Master.Protein.Accessions'
    }
    else{
      group_id <- 'Sequence'
    }
    
    contains_notch_1_2 <- x[,1:7] %>%
      get_notch_per_protein(master_prot_col=group_id) %>%
      group_by_at(group_id) %>%
      summarise(max_fraction=max(fraction_below)) %>%
      filter(max_fraction>0) %>%
      mutate(comparison='2 vs 1')
    
    contains_notch_1_6 <- x[,c(1:4, 8:10)] %>%
      get_notch_per_protein(master_prot_col=group_id) %>%
      group_by_at(group_id) %>%
      summarise(max_fraction=max(fraction_below)) %>%
      filter(max_fraction>0) %>%
      mutate(comparison='6 vs 1')
    
    return(rbind(contains_notch_1_2, contains_notch_1_6))
  })
})
```
Plot the precision and recall for the strict filtering +/- notch
```{r}
thresholds_to_retain <- all_limma_results %>%
    select(threshold) %>%
    unique() %>%
    separate(threshold, into=c('delta', 'interference', 'sn', 'notch'), sep=', ', remove=FALSE) %>%
    filter(sn==10, interference==10, delta==0.5) %>%
    pull(threshold)

to_plot <- all_limma_results %>%
  filter(level=='protein', run=='AGC: 5E4', threshold %in% thresholds_to_retain) %>%
  separate(threshold, into=c('delta', 'interference', 'sn', 'notch'), sep=', ', remove=FALSE) %>%
  mutate(notch=recode(notch, '1'=TRUE, '0'=FALSE))

keep_features <- to_plot %>%
  group_by(Row.names) %>%
  tally() %>%
  filter(n==4) %>% # 2 filtering threshold sets * 2 contrasts
  pull(Row.names)

to_plot_common <- to_plot %>%
  filter(Row.names %in% keep_features) %>%
  mutate(comparison=gsub('x', '', comparison)) %>%
  merge(contains_notch$`AGC: 5E4`$protein,
        by.x=c('comparison', 'Row.names'),
        by.y=c('comparison', 'Master.Protein.Accessions'), all.x=TRUE) %>%
  mutate(contains_notch=is.finite(max_fraction),
         sig=adj.P.Val<0.01,
         sig_up=(sig & logFC>0),
         sig_dw=(sig & logFC<0),
         tp=((sig_up & species=='S.cerevisiae') | (sig_dw & species=='H.sapiens')),
         fp=((sig_dw & species=='S.cerevisiae') | (sig_up & species=='H.sapiens')),
         fn=!sig)

summarised_all <- to_plot_common %>%
  group_by(threshold, comparison, species, delta, interference, sn, notch, run) %>%
  summarise(ave_lgfc=mean(logFC, na.rm=TRUE),
            tp=sum(tp),
            fp=sum(fp),
            fn=sum(fn),
            precision=tp/(tp+fp),
            recall=tp/(tp+fn)) %>%
  mutate(f1=2*((precision*recall/(precision+recall))),
         level='all')

summarised_with_notch <- to_plot_common %>%
  filter(contains_notch) %>%
  group_by(threshold, comparison, species, delta, interference, sn, notch, run) %>%
  summarise(ave_lgfc=mean(logFC, na.rm=TRUE),
            tp=sum(tp),
            fp=sum(fp),
            fn=sum(fn),
            precision=tp/(tp+fp),
            recall=tp/(tp+fn)) %>%
  mutate(f1=2*((precision*recall/(precision+recall))),
         level='notch')
```

```{r}
summarised <- rbind(summarised_all, summarised_with_notch) %>%
  ungroup() %>%
  select(notch, species, comparison, precision, recall, tp, fp, fn, level) %>%
  pivot_longer(cols=-c(notch, species, comparison, level), names_to='metric', values_to='value') %>%
  mutate(comp_spec=interaction(species, comparison, sep=': '))

summarised %>%
  #filter(!metric %in% c('tp', 'fp', 'fn')) %>%
  filter(!metric %in% c('precision', 'recall', 'fp')) %>%
  ggplot(aes(metric, value, fill=notch)) +
  geom_bar(stat='identity', position='dodge') +
  theme_camprot(base_size=10, border=FALSE) +
  facet_grid(level~comp_spec, scales='free_y') +
  scale_fill_manual(values=get_cat_palette(6)[5:6], name='Notch filtering') +
  #theme(axis.text.x=element_text(angle=45, vjust=1, hjust=1)) +
  ylab('Value') +
  xlab('') +
  geom_text(aes(label=round(value, 3), y=Inf), vjust=3, position = position_dodge(width = 1), size=2) +
  #scale_y_continuous(breaks=seq(0,3000,1000), limits=c(0,3000)) +
  theme(aspect.ratio=1) +
  scale_y_continuous(expand=expansion(mult = c(0, .2)))

summarised %>%
  filter(species=='H.sapiens', comparison=='2 vs 1', level=='notch', !metric %in% c('precision', 'recall', 'fp')) %>%
  mutate(metric=recode(metric, 'fn'='False negative', 'tp'='True positive')) %>%
  ggplot(aes(metric, value, fill=notch)) +
  geom_bar(stat='identity', position='dodge') +
  theme_camprot(border=FALSE) +
  scale_fill_manual(values=get_cat_palette(6)[5:6], name='Notch filtering') +
  #theme(axis.text.x=element_text(angle=45, vjust=1, hjust=1)) +
  ylab('Value') +
  xlab('') +
  geom_text(aes(label=round(value, 4), y=Inf), vjust=0.9, position = position_dodge(width = 1), size=4) +
  theme(aspect.ratio=1)

```

Plot precision, recall, F1 and median fold-change for each combination of input data and processing parameters.
```{r}

limma_filtering_sig_count_tile_plots <- runs %>%
  lapply(function(r){
    levels %>%
    lapply(function(l) {

    thresholds_to_retain <- all_limma_results %>%
      select(threshold) %>%
      unique() %>%
      separate(threshold, into=c('delta', 'interference', 'sn', 'notch'), sep=', ', remove=FALSE) %>%
      filter(sn %in% c(0, 10)) %>%
      pull(threshold)

    to_plot <- all_limma_results %>%
      filter(level==l, run==r, threshold %in% thresholds_to_retain) %>%
      separate(threshold, into=c('delta', 'interference', 'sn', 'notch'), sep=', ', remove=FALSE) %>%
      mutate(notch=recode(notch, '1'='With notch filt', '0'='Without notch filt.'),
             sn=factor(sn, levels=c(0,10,100)),
             interference=factor(interference, levels=c(10, 50, 100)))
    
    complete_feature_n <- to_plot %>% select(delta, interference, sn, notch) %>% unique() %>% nrow()
    
    to_plot %>%
      group_by(Row.names) %>%
      tally() %>%
      pull(n) %>%
      table()
    
    keep_features <- to_plot %>%
      group_by(Row.names) %>%
      tally() %>%
      filter(n==(complete_feature_n*2)) %>% # * 2 contrasts
      pull(Row.names)
    
    print(length(keep_features))

    summarised_common_features <- to_plot %>%
      filter(Row.names %in% keep_features) %>%
      mutate(comparison=gsub('x', '', comparison)) %>%
      mutate(sig=adj.P.Val<0.01,
             sig_up=(sig & logFC>0),
             sig_dw=(sig & logFC<0),
             tp=((sig_up & species=='S.cerevisiae') | (sig_dw & species=='H.sapiens')),
             fp=((sig_dw & species=='S.cerevisiae') | (sig_up & species=='H.sapiens')),
             fn=!sig) %>%
      group_by(threshold, comparison, species, delta, interference, sn, notch, run) %>%
      summarise(ave_lgfc=mean(logFC, na.rm=TRUE),
                tp=sum(tp),
                fp=sum(fp),
                fn=sum(fn),
                precision=tp/(tp+fp),
                recall=tp/(tp+fn)) %>%
      mutate(f1=2*((precision*recall/(precision+recall))))
    
    if(l=='protein'){
      group_id <- 'Master.Protein.Accessions'
    }
    else{
      group_id <- 'Sequence'
    }
    
    summarised_common_features_with_notch <- to_plot %>%
      filter(Row.names %in% keep_features) %>%
      mutate(comparison=gsub('x', '', comparison)) %>%
      merge(contains_notch[[r]][[l]],
            by.x=c('comparison', 'Row.names'),
            by.y=c('comparison', group_id)) %>%
      mutate(sig=adj.P.Val<0.01,
             sig_up=(sig & logFC>0),
             sig_dw=(sig & logFC<0),
             tp=((sig_up & species=='S.cerevisiae') | (sig_dw & species=='H.sapiens')),
             fp=((sig_dw & species=='S.cerevisiae') | (sig_up & species=='H.sapiens')),
             fn=!sig) %>%
      group_by(threshold, comparison, species, delta, interference, sn, notch, run) %>%
      summarise(ave_lgfc=median(logFC, na.rm=TRUE),
                tp=sum(tp),
                fp=sum(fp),
                fn=sum(fn),
                precision=tp/(tp+fp),
                recall=tp/(tp+fn)) %>%
      mutate(f1=2*((precision*recall/(precision+recall))))
          
    p <- summarised_common_features %>%
      ggplot(aes(y=interaction(interference, delta, sep=', '),
                 x=interaction(sn, notch, sep=', '),
                 fill=f1,
                 label=round(f1, 3))) +
      geom_tile() +
      theme_camprot(base_size=12) +
      facet_grid(comparison~species) +
      #ggtitle(sprintf('%s - %s', method, level)) +
      scale_fill_gradient2(mid='grey90', high=get_cat_palette(1),
                           low=get_cat_palette(2)[2],
                           midpoint=0.5,
                           name='F1 score') +
      ylab('Max interference (%), Min delta score') +
      xlab('Min Signal/Noise') +
      geom_text(size=3) +
      theme(panel.spacing = unit(0, "lines")) +
      scale_x_discrete(expand = c(0,0)) +
            scale_y_discrete(expand = c(0,0)) +
      theme(axis.text.x=element_text(angle=45, vjust=1, hjust=1)) +
      ggtitle(sprintf('%s - %s', r, l))
    
    print(p)
    print(p + aes(fill=precision, label=round(precision, 3)) +
          scale_fill_gradient2(mid='grey90', high=get_cat_palette(1),
                   low=get_cat_palette(2)[2],
                   midpoint=0.5,
                   name='Precision'))
    print(p + aes(fill=recall, label=round(recall, 3)) +
          scale_fill_gradient2(mid='grey90', high=get_cat_palette(1),
                   low=get_cat_palette(2)[2],
                   midpoint=0.5,
                   name='recall'))

    p2 <- p + aes(fill=2^ave_lgfc, label=round(2^ave_lgfc, 3)) +
                    scale_fill_gradient2(mid='grey90', high=get_cat_palette(1),
                           low=get_cat_palette(2)[2],
                           midpoint=0.5,
                           name='Median\nfold-change')
    print(p2)
    
    print(p %+% summarised_common_features_with_notch)
    print(p2 %+% summarised_common_features_with_notch)
    
    p3 <- to_plot %>%
      mutate(comparison=gsub('x', '', comparison)) %>%
      mutate(sig=adj.P.Val<0.01) %>%
      group_by(threshold, comparison, species, delta, interference, sn, notch, run) %>%
      summarise(n_sig=sum(as.numeric(sig), na.rm=TRUE)) %>%
      ggplot(aes(y=interaction(interference, delta, sep=', '),
                 x=interaction(sn, notch, sep=', '),
                 fill=n_sig,
                 label=n_sig)) +
      geom_tile() +
      theme_camprot(base_size=12) +
      facet_grid(comparison~species) +
      scale_fill_gradient2(mid='grey90', high=get_cat_palette(1),
                           low=get_cat_palette(2)[2],
                           midpoint=0.5,
                           name=sprintf('Significant\n%ss\n(FDR<1%%)', l)) +
      ylab('Max. interference (%), min. delta score') +
      xlab('Min Signal/Noise') +
      geom_text(size=3) +
      theme(panel.spacing = unit(0, "lines")) +
      scale_x_discrete(expand = c(0,0)) +
            scale_y_discrete(expand = c(0,0)) +
      theme(axis.text.x=element_text(angle=45, vjust=1, hjust=1)) +
      ggtitle(sprintf('%s - %s', r, l))
    
    print(p3)
    return(list('to_plot'=to_plot, 'keep_features'=keep_features, 'p'=p, 'p2'=p2, 'p3'=p3))
  })
})
```

Save out desired plots
```{r}
plots <- limma_filtering_sig_count_tile_plots$`AGC: 5E4`
for(level in names(plots)){
  ggsave(sprintf('../results/plots/%s_f1.png', level),
         (plots[[level]]$p + ggtitle('')), width=5, height=5)
  
  ggsave(sprintf('../results/plots/%s_fold_change.png', level),
         (plots[[level]]$p2 + ggtitle('')), width=5, height=5)
  
  ggsave(sprintf('../results/plots/%s_tally.png', level),
         (plots[[level]]$p3 + ggtitle('')), width=5, height=5)
}

```

Fold differences +/- notch filtering
```{r}
filtering_to_plot <- list('sn'=0, 'int'=10, delta=0.5)

for(level in names(limma_filtering_sig_count_tile_plots$`AGC: 5E4`)){
  
  to_plot <- limma_filtering_sig_count_tile_plots$`AGC: 5E4`[[level]]$to_plot %>%
    filter(sn==filtering_to_plot$sn,
           interference==filtering_to_plot$int,
           delta==filtering_to_plot$delta)
  
  print(dim(to_plot))

  p <- to_plot %>%
    mutate(comparison=gsub('x', '', comparison)) %>%
    ggplot(aes(logFC, colour=notch)) +
    theme_camprot(base_size=12) +
    facet_wrap(species~comparison, scales='free') +
    xlab('Fold difference (log2)') +
    ylab('Density') +
    scale_colour_manual(values=get_cat_palette(4)[3:4], name='') +
    geom_vline(data=expected[expected$comparison %in% c('2 vs 1', '6 vs 1'),],
               aes(xintercept=log2(expected)), linetype=2, lwd=1, colour='grey') +
    geom_density()
  
  print(p)
  ggsave(sprintf('../results/plots/%s_fold_change_distribution.png', level), p)

}


```
```{r}
psm_res_flt <- readRDS('../results/psm_res_flt.rds')
```

```{r}
thresholds_of_interest <- list(c('0.2, 10, 10, 0'), c('0.5, 10, 10, 0'))

contains_notch_filt <- psm_res_flt %>% lapply(function(run){
  contains_notch <- vector('list', length(thresholds_of_interest))
  names(contains_notch) <- thresholds_of_interest
  for(threshold in thresholds_of_interest){

    contains_notch_1_2 <- run[[threshold]][,1:7] %>%
      get_notch_per_protein() %>%
      group_by(Master.Protein.Accessions) %>%
      summarise(max_fraction=max(fraction_below)) %>%
      filter(max_fraction>0) %>%
      mutate(comparison='2 vs 1')
    
    contains_notch_1_6 <- run[[threshold]][,c(1:4, 8:10)] %>%
      get_notch_per_protein() %>%
      group_by(Master.Protein.Accessions) %>%
      summarise(max_fraction=max(fraction_below)) %>%
      filter(max_fraction>0) %>%
      mutate(comparison='6 vs 1')

    contains_notch[[threshold]] <- rbind(contains_notch_1_2, contains_notch_1_6)
  }
  contains_notch
})
```


```{r}

limma_filtering_sig_count_tile_plots$`AGC: 5E4`$protein$to_plot %>%
  filter(interference==50, sn==0, delta==0.2) %>%
  mutate(comparison=gsub('x', '', comparison)) %>%
  merge(contains_notch_filt$`AGC: 5E4`$`0.2, 10, 10, 0`,
        by.x=c('comparison', 'Row.names'),
        by.y=c('comparison', 'Master.Protein.Accessions')) %>%
  select(species, logFC, notch, interference, sn, comparison, Row.names) %>%
  ggplot(aes(logFC)) +
  geom_histogram() +
  theme_camprot(base_size=15) +
  facet_grid(cols=(vars(species, comparison)), rows=vars(notch), scales='free') +
  geom_vline(data=expected[expected$comparison %in% c('2 vs 1', '6 vs 1'),],
             aes(xintercept=log2(expected)), linetype=2, colour=get_cat_palette(1))
  

```


```{r}
limma_filtering_sig_count_tile_plots$`AGC: 5E4`$protein$to_plot %>%
  merge(contains_notch$`AGC: 5E4`$protein[,c('Master.Protein.Accessions', 'max_fraction')],
        by.x='Row.names', by.y='Master.Protein.Accessions') %>%
  filter(interference==10, sn==10, delta==0.5) %>%
  mutate(comparison=gsub('x', '', comparison)) %>%
  ggplot(aes(logFC, colour=notch)) +
  geom_density() +
  theme_camprot() +
  facet_grid(comparison~species, scales='free') +
  geom_vline(data=expected[expected$comparison %in% c('2 vs 1', '6 vs 1'),],
             aes(xintercept=log2(expected)))#, linetype==2, colour='grey')
  
```
```{r, fig.width=7, fig.height=7}
expected_data <- expected %>%
  filter(comparison %in% c('2 vs 1', '6 vs 1'), species=='S.cerevisiae') %>%
  mutate(`Without notch filt.`=log2(expected), `With notch filt`=log2(expected))

plus_minus_notch <- limma_filtering_sig_count_tile_plots$`AGC: 5E4`$protein$to_plot %>%
  filter(interference==10, sn==10, delta==0.5) %>%
  select(Row.names, logFC, species, comparison, level, run, notch) %>%
  pivot_wider(names_from=notch, values_from=logFC) %>%
  mutate(comparison=gsub('x', '', comparison)) %>%
  merge(contains_notch_filt$`AGC: 5E4`$`0.5, 10, 10, 0`,
        by.x=c('Row.names', 'comparison'), by.y=c('Master.Protein.Accessions', 'comparison'))

plus_minus_notch %>%
  filter(species=='S.cerevisiae') %>%
  ggplot(aes(`Without notch filt.`, `With notch filt`)) +
  geom_abline(slope=1, linetype=2) +
  geom_point() +
  theme_camprot(base_size=10) +
  facet_wrap(~comparison, scales='free') +
  xlab('Without notch filtering') +
  ylab('With notch filtering') +
  geom_point(data=expected_data, colour=get_cat_palette(5)[5], size=3)
```
How does notch filtering alter fold changes relative to the ground truth.
```{r}
library(ggbeeswarm)

change_relative_to_gt <- plus_minus_notch  %>%
  filter(is.finite(`With notch filt`)) %>%
  merge(expected[expected$comparison %in% c('2 vs 1', '6 vs 1'),], by=c('species', 'comparison')) %>%
  mutate(expected=log2(expected)) %>%
  mutate(without_diff=abs(`Without notch filt.`-expected),
         with_diff=abs(`With notch filt`-expected),
         improvement=without_diff-with_diff) %>% arrange(improvement)

table(change_relative_to_gt$species)
dim(change_relative_to_gt)

p <- change_relative_to_gt %>% 
  filter(species=='S.cerevisiae') %>%
  ggplot(aes(comparison, improvement)) +
  theme_camprot(base_size=15) +
  xlab('Comparison') +
  ylab('Change in fold-change with notch-filtering,\nrelative to ground-truth') +
  theme(aspect.ratio=2) +
  ylim(-0.2,0.2) +
  geom_quasirandom()

print(p)
ggsave('../results/plots/fold_change_plus_minus_notch_filt.png')
```
